Version 4.0.2-02 Refactoring - change type of 'count' to double

+++ src/MainWindow.java
@@ -87,7 +87,7 @@ public class MainWindow extends JFrame {
             }
         });
 
-        int count = flag ? 10 : 10;
+        double count = flag ? 10.0 : 10.0;
